Fixed accumulators in gemm for LNS16 being incorrectly initialized
cutlass has an Array class, which is essentially std::vector that also works on GPU. There is a specialization of this class for types that are less that 32 bits in size, since the size of GPU registers is 32 bits. This specialization stores the data in uint32_t (two lns per uint32_t) for arrays with even number of elements, and in uint16_t for arrays with odd number of elements. The clear() method of this specialization simply sets the memory to 0, but since it stores lns, it actually sets all of them to 1.
The solution was to specialize the Array class for LNS16 and change the clear() method.